[FIX] purchase, purchase_stock: replace test in the correct module
The tests were placed in `purchase` module but as using 'detailed_type': 'product' adds product in stock so tests should be placed in purchase_stock instead

Related PR: https://github.com/odoo/odoo/pull/145440
Related fw-port: https://github.com/odoo/odoo/pull/146814
