+++ addons/purchase/tests/test_purchase.py
@@ -367,27 +367,3 @@ class TestPurchase(AccountTestInvoicingCommon):
         po.button_confirm()
 
         self.assertEqual(po.order_line.product_id.seller_ids.mapped('name'), delivery_address)
-
-    def test_product_purchase_method_with_receive_as_default_purchase_method(self):
-        self.env['ir.default'].set('product.template', 'purchase_method', 'receive', company_id=True)
-
-        product = self.env['product.product'].create({'name': 'product_test'})
-        self.assertEqual(product.purchase_method, 'receive')
-
-        product.update({'detailed_type': 'service'})
-        self.assertEqual(product.purchase_method, 'purchase')
-
-        product.update({'detailed_type': 'product'})
-        self.assertEqual(product.purchase_method, 'receive')
-
-    def test_product_purchase_method_with_purchase_as_default_purchase_method(self):
-        self.env['ir.default'].set('product.template', 'purchase_method', 'purchase', company_id=True)
-
-        product = self.env['product.product'].create({'name': 'product_test'})
-        self.assertEqual(product.purchase_method, 'purchase')
-
-        product.update({'detailed_type': 'service'})
-        self.assertEqual(product.purchase_method, 'purchase')
-
-        product.update({'detailed_type': 'product'})
-        self.assertEqual(product.purchase_method, 'purchase')
