+++ addons/purchase_stock/tests/__init__.py
@@ -19,3 +19,4 @@ from . import test_move_cancel_propagation
 from . import test_product_template
 from . import test_routes
 from . import test_purchase_stock_accrued_entries
+from . import test_purchase_method
